Reorganized repo based on modules of related content. New modules:  json, pandas
