+++ src/unifair/__init__.py
@@ -1,1 +1,5 @@
 __version__ = '0.1.0'
+
+import os
+
+ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
